Added Map vents, cameras, sensors Fixed translations
