+++ src/utils/AUMT/Scenes/Maps/MiraHQ.ts
@@ -7,18 +7,19 @@ import {
 } from "constants/maps";
 
 import Layer from "utils/AUMT/Layer";
-import Map from "../Map";
+import MapScene from "../MapScene";
 import TextLine from "utils/AUMT/Entities/TextLine";
 import Vector from "utils/math/Vector";
 import i18n from "utils/i18n";
 
-export default class MiraHQ extends Map {
+export default class MiraHQ extends MapScene {
   public constructor() {
     super("MiraHQ", MIRAHQ_SCALE);
 
     const areas = new Layer();
 
-    areas.entities.push(
+    this.texts.set(
+      "greenhouse",
       new TextLine(
         i18n.t("maps.greenhouse"),
         new Vector(8000, 1900),
@@ -30,7 +31,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "office",
       new TextLine(
         i18n.t("maps.office"),
         new Vector(7200, 3300),
@@ -42,7 +44,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "admin",
       new TextLine(
         i18n.t("maps.admin"),
         new Vector(8750, 3300),
@@ -54,7 +57,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "reactor",
       new TextLine(
         i18n.t("maps.reactor"),
         new Vector(4000, 5200),
@@ -66,7 +70,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "laboratory",
       new TextLine(
         i18n.t("maps.laboratory"),
         new Vector(5800, 5200),
@@ -78,7 +83,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "decontamination",
       new TextLine(
         i18n.t("maps.decontamination"),
         new Vector(4900, 6000),
@@ -90,7 +96,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "launchpad",
       new TextLine(
         i18n.t("maps.launchpad"),
         new Vector(2000, 7100),
@@ -102,7 +109,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "lockerRoom",
       new TextLine(
         i18n.t("maps.lockerRoom"),
         new Vector(5900, 6800),
@@ -114,7 +122,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "communications",
       new TextLine(
         i18n.t("maps.communications"),
         new Vector(7300, 7100),
@@ -126,7 +135,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "medBay",
       new TextLine(
         i18n.t("maps.medBay"),
         new Vector(7300, 8000),
@@ -138,7 +148,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "storage",
       new TextLine(
         i18n.t("maps.storage"),
         new Vector(8425, 6650),
@@ -150,7 +161,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "cafeteria",
       new TextLine(
         i18n.t("maps.cafeteria"),
         new Vector(10000, 6650),
@@ -162,7 +174,8 @@ export default class MiraHQ extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "balcony",
       new TextLine(
         i18n.t("maps.balcony"),
         new Vector(9550, 8450),
@@ -174,11 +187,9 @@ export default class MiraHQ extends Map {
       )
     );
 
-    this.layers.push(areas);
+    this.texts.forEach((value) => areas.entities.push(value));
 
-    // TODO - Add tasks
-    // TODO - Add zones
-    // TODO - Add vents/vent connections
+    this.layers.push(areas);
 
     this.layers.push(this.players);
   }
