+++ src/utils/AUMT/Scenes/Maps/Polus.ts
@@ -7,18 +7,19 @@ import {
 } from "constants/maps";
 
 import Layer from "utils/AUMT/Layer";
-import Map from "../Map";
+import MapScene from "../MapScene";
 import TextLine from "utils/AUMT/Entities/TextLine";
 import Vector from "utils/math/Vector";
 import i18n from "utils/i18n";
 
-export default class Polus extends Map {
+export default class Polus extends MapScene {
   public constructor() {
     super("Polus", POLUS_SCALE);
 
     const areas = new Layer();
 
-    areas.entities.push(
+    this.texts.set(
+      "dropship",
       new TextLine(
         i18n.t("maps.dropship"),
         new Vector(4600, 1000),
@@ -30,7 +31,8 @@ export default class Polus extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "security",
       new TextLine(
         i18n.t("maps.security"),
         new Vector(950, 3700),
@@ -42,7 +44,8 @@ export default class Polus extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "electrical",
       new TextLine(
         i18n.t("maps.electrical"),
         new Vector(2600, 3850),
@@ -54,7 +57,8 @@ export default class Polus extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "storage",
       new TextLine(
         i18n.t("maps.storage"),
         new Vector(5600, 3800),
@@ -66,7 +70,8 @@ export default class Polus extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "laboratory",
       new TextLine(
         i18n.t("maps.laboratory"),
         new Vector(9400, 2800),
@@ -78,7 +83,8 @@ export default class Polus extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "o2",
       new TextLine(
         i18n.t("maps.o2"),
         new Vector(800, 5200),
@@ -90,7 +96,8 @@ export default class Polus extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "communications",
       new TextLine(
         i18n.t("maps.communications"),
         new Vector(3200, 5000),
@@ -102,7 +109,8 @@ export default class Polus extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "office",
       new TextLine(
         i18n.t("maps.office"),
         new Vector(5400, 5300),
@@ -114,7 +122,8 @@ export default class Polus extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "weapons",
       new TextLine(
         i18n.t("maps.weapons"),
         new Vector(3400, 6800),
@@ -126,7 +135,8 @@ export default class Polus extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "admin",
       new TextLine(
         i18n.t("maps.admin"),
         new Vector(6400, 6600),
@@ -138,7 +148,8 @@ export default class Polus extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "specimenRoom",
       new TextLine(
         i18n.t("maps.specimenRoom"),
         new Vector(9800, 6000),
@@ -150,11 +161,9 @@ export default class Polus extends Map {
       )
     );
 
-    this.layers.push(areas);
+    this.texts.forEach((value) => areas.entities.push(value));
 
-    // TODO - Add tasks
-    // TODO - Add zones
-    // TODO - Add vents/vent connections
+    this.layers.push(areas);
 
     this.layers.push(this.players);
   }
