+++ src/utils/AUMT/Scenes/Maps/TheAirship.ts
@@ -7,19 +7,19 @@ import {
 } from "constants/maps";
 
 import Layer from "utils/AUMT/Layer";
-import Line from "utils/AUMT/Entities/Line";
-import Map from "../Map";
+import MapScene from "../MapScene";
 import TextLine from "utils/AUMT/Entities/TextLine";
 import Vector from "utils/math/Vector";
 import i18n from "utils/i18n";
 
-export default class TheAirship extends Map {
+export default class TheAirship extends MapScene {
   public constructor() {
     super("TheAirship", THE_AIRSHIP_SCALE);
 
     const areas = new Layer();
 
-    areas.entities.push(
+    this.texts.set(
+      "meetingRoom",
       new TextLine(
         i18n.t("maps.meetingRoom"),
         new Vector(7800, 800),
@@ -31,7 +31,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "vault",
       new TextLine(
         i18n.t("maps.vault"),
         new Vector(3500, 1200),
@@ -43,7 +44,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "brig",
       new TextLine(
         i18n.t("maps.brig"),
         new Vector(5300, 2000),
@@ -55,7 +57,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "gapRoom",
       new TextLine(
         i18n.t("maps.gapRoom"),
         new Vector(7000, 2000),
@@ -67,7 +70,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "records",
       new TextLine(
         i18n.t("maps.records"),
         new Vector(9700, 1600),
@@ -79,7 +83,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "lounge",
       new TextLine(
         i18n.t("maps.lounge"),
         new Vector(11400, 2700),
@@ -91,7 +96,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "cockpit",
       new TextLine(
         i18n.t("maps.cockpit"),
         new Vector(1000, 3800),
@@ -103,7 +109,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "communications",
       new TextLine(
         i18n.t("maps.communications"),
         new Vector(2600, 3600),
@@ -115,7 +122,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "engineRoom",
       new TextLine(
         i18n.t("maps.engineRoom"),
         new Vector(5000, 4100),
@@ -127,7 +135,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "mainHall",
       new TextLine(
         i18n.t("maps.mainHall"),
         new Vector(7800, 3900),
@@ -139,7 +148,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "showers",
       new TextLine(
         i18n.t("maps.showers"),
         new Vector(10000, 3800),
@@ -151,7 +161,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "cargoBay",
       new TextLine(
         i18n.t("maps.cargoBay"),
         new Vector(13000, 4100),
@@ -163,7 +174,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "armory",
       new TextLine(
         i18n.t("maps.armory"),
         new Vector(2800, 5200),
@@ -175,7 +187,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "viewingDeck",
       new TextLine(
         i18n.t("maps.viewingDeck"),
         new Vector(2600, 6600),
@@ -187,7 +200,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "kitchen",
       new TextLine(
         i18n.t("maps.kitchen"),
         new Vector(4400, 6300),
@@ -199,7 +213,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "security",
       new TextLine(
         i18n.t("maps.security"),
         new Vector(6900, 6600),
@@ -211,7 +226,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "electrical",
       new TextLine(
         i18n.t("maps.electrical"),
         new Vector(8900, 5800),
@@ -223,7 +239,8 @@ export default class TheAirship extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "medical",
       new TextLine(
         i18n.t("maps.medical"),
         new Vector(10600, 6000),
@@ -235,10 +252,9 @@ export default class TheAirship extends Map {
       )
     );
 
+    this.texts.forEach((value) => areas.entities.push(value));
+
     this.layers.push(areas);
-    // TODO - Add areas
-    // TODO - Add tasks
-    // TODO - Add zones
 
     this.layers.push(this.players);
   }
