+++ src/utils/AUMT/Scenes/Maps/TheSkeld.ts
@@ -7,18 +7,19 @@ import {
 } from "constants/maps";
 
 import Layer from "utils/AUMT/Layer";
-import Map from "../Map";
+import MapScene from "../MapScene";
 import TextLine from "utils/AUMT/Entities/TextLine";
 import Vector from "utils/math/Vector";
 import i18n from "utils/i18n";
 
-export default class TheSkeld extends Map {
+export default class TheSkeld extends MapScene {
   public constructor() {
     super("TheSkeld", THE_SKELD_SCALE);
 
     const areas = new Layer();
 
-    areas.entities.push(
+    this.texts.set(
+      "upperEngine",
       new TextLine(
         i18n.t("maps.upperEngine"),
         new Vector(1800, 1400),
@@ -30,7 +31,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "cafeteria",
       new TextLine(
         i18n.t("maps.cafeteria"),
         new Vector(4800, 1400),
@@ -42,7 +44,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "weapons",
       new TextLine(
         i18n.t("maps.weapons"),
         new Vector(6600, 1100),
@@ -54,7 +57,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "reactor",
       new TextLine(
         i18n.t("maps.reactor"),
         new Vector(1200, 2200),
@@ -66,7 +70,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "security",
       new TextLine(
         i18n.t("maps.security"),
         new Vector(2600, 2200),
@@ -78,7 +83,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "medBay",
       new TextLine(
         i18n.t("maps.medBay"),
         new Vector(3400, 1800),
@@ -90,7 +96,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "o2",
       new TextLine(
         i18n.t("maps.o2"),
         new Vector(6200, 1900),
@@ -102,7 +109,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "navigation",
       new TextLine(
         i18n.t("maps.navigation"),
         new Vector(7600, 2100),
@@ -114,7 +122,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "lowerEngine",
       new TextLine(
         i18n.t("maps.lowerEngine"),
         new Vector(1800, 3000),
@@ -126,7 +135,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "electrical",
       new TextLine(
         i18n.t("maps.electrical"),
         new Vector(3600, 2800),
@@ -138,7 +148,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "storage",
       new TextLine(
         i18n.t("maps.storage"),
         new Vector(4700, 3300),
@@ -150,7 +161,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "admin",
       new TextLine(
         i18n.t("maps.admin"),
         new Vector(5800, 3000),
@@ -162,7 +174,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "communications",
       new TextLine(
         i18n.t("maps.communications"),
         new Vector(5800, 4000),
@@ -174,7 +187,8 @@ export default class TheSkeld extends Map {
       )
     );
 
-    areas.entities.push(
+    this.texts.set(
+      "shields",
       new TextLine(
         i18n.t("maps.shields"),
         new Vector(6400, 3500),
@@ -186,10 +200,9 @@ export default class TheSkeld extends Map {
       )
     );
 
+    this.texts.forEach((value) => areas.entities.push(value));
+
     this.layers.push(areas);
-    // TODO - Add tasks
-    // TODO - Add zones
-    // TODO - Add vents/vent connections
 
     this.layers.push(this.players);
   }
