+++ src/utils/AUMT/Scenes/Menu.ts
@@ -8,6 +8,7 @@ import SceneManager from "../SceneManager";
 import Sprite from "../Entities/Sprite";
 import TextLine from "../Entities/TextLine";
 import Vector from "utils/math/Vector";
+import i18n from "utils/i18n";
 
 export default class Menu extends Scene {
   public constructor() {
@@ -15,9 +16,8 @@ export default class Menu extends Scene {
 
     const buttons = new Layer();
 
-    // TODO - Translate
     buttons.entities.push(
-      new TextLine("Choose a map:", new Vector(960, 400), true, 60)
+      new TextLine(i18n.t("maps.chooseMap"), new Vector(960, 400), true, 60)
     );
 
     buttons.entities.push(
@@ -78,6 +78,10 @@ export default class Menu extends Scene {
 
   public onExit(): void {
     // Nothing to do here
+  }
+
+  public updateText(): void {
+    (<TextLine>this.layers[0].entities[0]).setText(i18n.t("maps.chooseMap"));
   }
 
   public update(): void {
